Actually fix omgwtfnzbs size...

+++ nzbhydra/searchmodules/omgwtf.py
@@ -231,7 +231,7 @@ class OmgWtf(SearchModule):
                 entry.guid = item.find("nzbid").text
                 entry.title = item.find("release").text
                 entry.group = item.find("group").text
-                entry.size = item.find("sizebytes").text
+                entry.size = long(item.find("sizebytes").text)
                 entry.epoch = long(item.find("usenetage").text)
                 pubdate = arrow.get(entry.epoch)
                 entry.pubdate_utc = pubdate.format("ddd, DD MMM YYYY HH:mm:ss Z")
